spark-notebook/requirements.txt
boto>=2.45.0
boto3>=1.4.0
botocore>=1.5.7
Flask>=0.12
future>=0.16.0
PyYAML>=3.12
requests>=2.20.0